2022/2023 Car Point Autos Marketing Campaign.md
Car Point, a company that deals with the sales of cars, embarked on a marketing campaign that lasted the entire 4th quarter of 2022. The company is interested in knowing the sort of effect the marketing campaign had in their sales, most especially in the year 2023.
A dataset containing the daily transactions of the company from January 2022 to December 2023 have been provided.